+++ web/app/themes/rebe/src/assets/scripts/constants.js
@@ -1,0 +1,3 @@
+angular.module('constants',[])
+  .constant('apiUrl', '/api/')
+  .constant('templatesPath', '/app/themes/rebe/src/views/ng-apps/');