+++ web/app/themes/rebe/src/assets/scripts/events/directives/preview.directive.js
@@ -1,0 +1,21 @@
+angular
+  .module('events.preview.directive', ['ui.router'])
+  .directive('eventPreview', function () {
+    return {
+      restrict: 'E',
+      replace: true,
+      templateUrl: '/app/themes/rebe/src/views/ng-apps/events/preview.html',
+      scope: {
+        image: '@',
+        title: '@',
+        year: '@',
+        slug: '@'
+      },
+      link: function (scope, element, attr) {
+        var imgArr = attr.image.split('.');
+        imgArr[imgArr.length-2] += '-250x250';
+        var previewImgPath = imgArr.join('.');
+        element.css('background-image', 'url('+ previewImgPath +')');
+      }
+    };
+  });