+++ web/app/themes/rebe/src/assets/scripts/stories/directives/preview.directive.js
@@ -1,0 +1,12 @@
+angular
+  .module('stories.preview.directive', ['ui.router'])
+  .directive('storyPreview', function () {
+    return {
+      restrict: 'E',
+      replace: true,
+      templateUrl:  '/app/themes/rebe/src/views/ng-apps/stories/preview.html',
+      link: function (scope, element) {
+        // 
+      }
+    };
+  });